% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSimu.R
\name{scSimu}
\alias{scSimu}
\title{Simulate drop-seq scRNA count data}
\usage{
scSimu(
  mu,
  alpha,
  count_dat = NULL,
  gene_name = NULL,
  cell_name = NULL,
  seq_depth = NULL,
  IND = T,
  cor_mat = NULL,
  cor_gene = NULL,
  sig_level = 0.05,
  strength_level = 0,
  seed = 7232024
)
}
\arguments{
\item{mu}{A vector contains gene mean expression level.}

\item{alpha}{A vector contains gene dispersion parameter (Shape parameter in the gamma distribution).}

\item{count_dat}{A gene x cell count matrix that user wants to mimic. If a count matrix is provided, then gene name, cell name and sequencing depth will be calculated automatically.}

\item{gene_name}{A vector contains gene names.}

\item{cell_name}{A vector contains cell name.}

\item{seq_depth}{A vector contains cell sequencing depth.}

\item{IND}{Simulate independent genes or not.}

\item{cor_mat}{A correlation matrix for simulating correlated genes. If count matrix is not provided, this is required for simulating correlated genes.}

\item{cor_gene}{User can also provides a vector of gene names if the count matrix is provided. scSimu will automatically estimate the correlation matrix for these genes.}

\item{sig_level}{The p-value threshold if a vector of gene names is provided to estimate the correlation matrix.}

\item{strength_level}{The correlation strength threshold if a vector of gene names is provided to estimate the correlation matrix.}

\item{seed}{Random seed}
}
\value{
A simulated count data
}
\description{
This function simulates Drop-seq scRNA count data using negative binomial distribution.
The input can be a count matrix with gene mean expression level and dispersion parameter or user defined parameters.
}
