% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_fit.R
\name{fit_smooth}
\alias{fit_smooth}
\title{Fit a smooth curve between mu and alpha}
\usage{
fit_smooth(marginal_para)
}
\arguments{
\item{marginal_para}{A data frame where}
}
\value{
A data.frame containing:
\itemize{
  \item \code{gene}: Gene name
  \item \code{mu}: The mean expression level of gene on log10 scale
  \item \code{alpha}: The shape parameter of gene for the Gamma distribution on log10 scale
  \item \code{deviances}: Deviance of the fit
}
}
\description{
This function fits a kernel regression smooth curve between mu and alpha.
}
