% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_fit.R
\name{marginal_fit}
\alias{marginal_fit}
\title{Estimate the marginal parameters of genes}
\usage{
marginal_fit(ori_ct)
}
\arguments{
\item{ori_ct}{Gene-by-cell count matrix with genes in rows and cells in columns.}
}
\value{
A data.frame containing:
\itemize{
  \item \code{gene}: Gene name
  \item \code{mu}: The mean expression level of gene on log10 scale
  \item \code{alpha}: The shape parameter of gene for the Gamma distribution on log10 scale
  \item \code{deviances}: Deviance of the fit
}
}
\description{
This function estimates the marginal parameters of genes by using the glm_gp()
function in glmGamPoi package.
}
